requirements.txt
Image==1.5.32

opencv-python~=4.2.0.34
numpy~=1.18.5
Pillow~=7.1.2
keyboard~=0.13.5
pandas~=1.0.4

tensorflow==2.2.0
pip~=19.2.3
wheel~=0.34.2
rsa~=4.6
pyasn1~=0.4.8
six~=1.15.0
gast~=0.3.3
h5py~=2.10.0
pytz~=2020.1
Django~=3.0.7
requests~=2.23.0
setuptools~=41.2.0
scipy~=1.4.1
matplotlib~=3.2.2
sqlparse~=0.3.1
asgiref~=3.2.9
cachetools~=4.1.0
certifi~=2020.4.5.2
urllib3~=1.25.9
python-dateutil~=2.8.1
chardet~=3.0.4
Markdown~=3.2.2
oauthlib~=3.1.0
idna~=2.9
Werkzeug~=1.0.1
tensorboard~=2.2.2
wrapt~=1.12.1
astunparse~=1.6.3
termcolor~=1.1.0
scikit-image~=0.17.2
mpylayer~=0.2a1
pytesseract==0.3.4
google-cloud-vision==0.25.1
google-auth<2.0dev,>=1.18.0
future~=0.16.0
dispatch~=1.3.1
astor~=0.8.1
google-api-python-client==1.9.3
pygdrive3==0.7.0
apiclient==1.0.4
oauth2client==4.1.3
